Move platform layer to src/platform/linux/ and split into modules
Restructure per impl_plan.md to support future platform targets
(windows, mac). Split monolithic platform_linux.c into:
  - platform_linux.c: callbacks, event handlers, main entry point
  - routing.c: ip route split tunnel setup/cleanup
  - killswitch.c: iptables/ip6tables kill switch rules
  - platform_internal.h: shared platform_ctx_t definition
