+++ src/platform/linux/platform_linux.c
@@ -5,17 +5,13 @@
  *   - libevent event loop driving tick()
  *   - TUN device creation and I/O
  *   - UDP socket creation via path_mgr
- *   - Split tunneling (routing table)
- *   - Kill switch (iptables)
- *   - DNS override (resolv.conf)
  *   - Signal handling (SIGINT/SIGTERM)
+ *
+ * Routing and killswitch are in separate files (routing.c, killswitch.c).
  */
 
+#include "platform_internal.h"
 #include "platform_linux.h"
-#include "libmqvpn.h"
-#include "tun.h"
-#include "dns.h"
-#include "path_mgr.h"
 #include "log.h"
 
 #include <stdio.h>
@@ -24,373 +20,6 @@
 #include <unistd.h>
 #include <errno.h>
 #include <signal.h>
-#include <sys/wait.h>
-#include <inttypes.h>
-#include <arpa/inet.h>
-#include <net/if.h>
-
-#include <event2/event.h>
-
-/* ─── Platform context ─── */
-
-typedef struct {
-    mqvpn_client_t     *client;
-
-    /* Event loop */
-    struct event_base  *eb;
-    struct event       *ev_tick;
-    struct event       *ev_tun;
-    struct event       *ev_sigint;
-    struct event       *ev_sigterm;
-
-    /* Path manager (UDP sockets) */
-    mqvpn_path_mgr_t    path_mgr;
-    mqvpn_path_handle_t  lib_path_handles[MQVPN_MAX_PATHS];
-    struct event        *ev_udp[MQVPN_MAX_PATHS];
-
-    /* TUN device */
-    mqvpn_tun_t          tun;
-    int                  tun_up;
-
-    /* Server address */
-    struct sockaddr_storage server_addr;
-    socklen_t            server_addrlen;
-
-    /* Split tunneling state */
-    int                  routing_configured;
-    int                  routing6_configured;
-    char                 orig_gateway[INET6_ADDRSTRLEN];
-    char                 orig_iface[IFNAMSIZ];
-    char                 server_ip_str[INET6_ADDRSTRLEN];
-    int                  server_port;
-    int                  has_v6;
-
-    /* DNS */
-    mqvpn_dns_t          dns;
-
-    /* Kill switch */
-    int                  killswitch_active;
-    int                  killswitch_enabled;
-    char                 ks_comment[64];
-
-    /* Shutdown */
-    int                  shutting_down;
-} platform_ctx_t;
-
-/* ================================================================
- *  Helper: run external commands
- * ================================================================ */
-
-static int
-run_ip_cmd(const char *const argv[])
-{
-    pid_t pid = fork();
-    if (pid < 0) return -1;
-    if (pid == 0) {
-        execvp("ip", (char *const *)argv);
-        _exit(127);
-    }
-    int status = 0;
-    while (waitpid(pid, &status, 0) < 0)
-        if (errno != EINTR) return -1;
-    return (WIFEXITED(status) && WEXITSTATUS(status) == 0) ? 0 : -1;
-}
-
-static int
-run_iptables_cmd(const char *const argv[])
-{
-    pid_t pid = fork();
-    if (pid < 0) return -1;
-    if (pid == 0) {
-        execvp(argv[0], (char *const *)argv);
-        _exit(127);
-    }
-    int status = 0;
-    while (waitpid(pid, &status, 0) < 0)
-        if (errno != EINTR) return -1;
-    return (WIFEXITED(status) && WEXITSTATUS(status) == 0) ? 0 : -1;
-}
-
-/* ================================================================
- *  Route discovery
- * ================================================================ */
-
-static int
-discover_route(const char *server_ip, sa_family_t af,
-               char *gateway, size_t gw_len,
-               char *iface, size_t if_len)
-{
-    int fds[2];
-    if (pipe(fds) < 0) return -1;
-
-    pid_t pid = fork();
-    if (pid < 0) { close(fds[0]); close(fds[1]); return -1; }
-
-    if (pid == 0) {
-        const char *const a4[] = {"ip", "-4", "route", "get", server_ip, NULL};
-        const char *const a6[] = {"ip", "-6", "route", "get", server_ip, NULL};
-        close(fds[0]);
-        if (dup2(fds[1], STDOUT_FILENO) < 0) _exit(127);
-        close(fds[1]);
-        execvp("ip", (char *const *)((af == AF_INET6) ? a6 : a4));
-        _exit(127);
-    }
-
-    close(fds[1]);
-    char out[1024];
-    ssize_t nread = read(fds[0], out, sizeof(out) - 1);
-    close(fds[0]);
-
-    int status = 0;
-    while (waitpid(pid, &status, 0) < 0)
-        if (errno != EINTR) return -1;
-    if (!WIFEXITED(status) || WEXITSTATUS(status) != 0 || nread <= 0) return -1;
-
-    out[nread] = '\0';
-    gateway[0] = '\0';
-    iface[0] = '\0';
-
-    char *saveptr = NULL;
-    for (char *tok = strtok_r(out, " \t\r\n", &saveptr); tok;
-         tok = strtok_r(NULL, " \t\r\n", &saveptr)) {
-        if (strcmp(tok, "via") == 0) {
-            tok = strtok_r(NULL, " \t\r\n", &saveptr);
-            if (tok) snprintf(gateway, gw_len, "%s", tok);
-        } else if (strcmp(tok, "dev") == 0) {
-            tok = strtok_r(NULL, " \t\r\n", &saveptr);
-            if (tok) snprintf(iface, if_len, "%s", tok);
-        }
-    }
-    return iface[0] ? 0 : -1;
-}
-
-/* ================================================================
- *  Routing setup / cleanup
- * ================================================================ */
-
-static int
-setup_routes(platform_ctx_t *p)
-{
-    sa_family_t af = p->server_addr.ss_family;
-    int prefix = mqvpn_sa_host_prefix(&p->server_addr);
-    mqvpn_sa_ntop(&p->server_addr, p->server_ip_str, sizeof(p->server_ip_str));
-
-    if (discover_route(p->server_ip_str, af,
-                       p->orig_gateway, sizeof(p->orig_gateway),
-                       p->orig_iface, sizeof(p->orig_iface)) < 0) {
-        LOG_WRN("could not determine original iface for %s", p->server_ip_str);
-        return -1;
-    }
-
-    char host_cidr[INET6_ADDRSTRLEN + 5];
-    snprintf(host_cidr, sizeof(host_cidr), "%s/%d", p->server_ip_str, prefix);
-    const char *ip_flag = (af == AF_INET6) ? "-6" : "-4";
-
-    if (p->orig_gateway[0] != '\0') {
-        LOG_INF("split tunnel: server %s via %s dev %s",
-                p->server_ip_str, p->orig_gateway, p->orig_iface);
-        const char *const pin[] = {
-            "ip", ip_flag, "route", "replace", host_cidr, "via", p->orig_gateway,
-            "dev", p->orig_iface, NULL
-        };
-        if (run_ip_cmd(pin) < 0) {
-            LOG_WRN("failed to pin server route");
-            return -1;
-        }
-    } else {
-        LOG_INF("split tunnel: server %s on-link dev %s", p->server_ip_str, p->orig_iface);
-    }
-
-    const char *const low[]  = {"ip", "route", "replace", "0.0.0.0/1", "dev", p->tun.name, NULL};
-    const char *const high[] = {"ip", "route", "replace", "128.0.0.0/1", "dev", p->tun.name, NULL};
-    if (run_ip_cmd(low) < 0 || run_ip_cmd(high) < 0) {
-        LOG_WRN("failed to set catch-all routes via %s", p->tun.name);
-        const char *u1[] = {"ip", "route", "del", "0.0.0.0/1", "dev", p->tun.name, NULL};
-        const char *u2[] = {"ip", "route", "del", "128.0.0.0/1", "dev", p->tun.name, NULL};
-        (void)run_ip_cmd(u1); (void)run_ip_cmd(u2);
-        if (p->orig_gateway[0]) {
-            const char *u3[] = {"ip", ip_flag, "route", "del", host_cidr,
-                                "via", p->orig_gateway, "dev", p->orig_iface, NULL};
-            (void)run_ip_cmd(u3);
-        }
-        return -1;
-    }
-    p->routing_configured = 1;
-
-    /* IPv6 catch-all routes */
-    if (p->has_v6) {
-        const char *v6l[] = {"ip", "-6", "route", "replace", "::/1", "dev", p->tun.name, NULL};
-        const char *v6h[] = {"ip", "-6", "route", "replace", "8000::/1", "dev", p->tun.name, NULL};
-        if (run_ip_cmd(v6l) == 0 && run_ip_cmd(v6h) == 0) {
-            p->routing6_configured = 1;
-            LOG_INF("IPv6 catch-all routes set via %s", p->tun.name);
-        } else {
-            LOG_WRN("failed to set IPv6 catch-all routes (continuing IPv4-only)");
-        }
-    }
-    return 0;
-}
-
-static void
-cleanup_routes(platform_ctx_t *p)
-{
-    if (!p->routing_configured) return;
-
-    if (p->routing6_configured) {
-        const char *d1[] = {"ip", "-6", "route", "del", "::/1", "dev", p->tun.name, NULL};
-        const char *d2[] = {"ip", "-6", "route", "del", "8000::/1", "dev", p->tun.name, NULL};
-        (void)run_ip_cmd(d1); (void)run_ip_cmd(d2);
-        p->routing6_configured = 0;
-    }
-
-    const char *d3[] = {"ip", "route", "del", "0.0.0.0/1", "dev", p->tun.name, NULL};
-    const char *d4[] = {"ip", "route", "del", "128.0.0.0/1", "dev", p->tun.name, NULL};
-    (void)run_ip_cmd(d3); (void)run_ip_cmd(d4);
-
-    if (p->orig_gateway[0]) {
-        const char *fl = (p->server_addr.ss_family == AF_INET6) ? "-6" : "-4";
-        int pfx = mqvpn_sa_host_prefix(&p->server_addr);
-        char hc[INET6_ADDRSTRLEN + 5];
-        snprintf(hc, sizeof(hc), "%s/%d", p->server_ip_str, pfx);
-        const char *d5[] = {"ip", fl, "route", "del", hc,
-                            "via", p->orig_gateway, "dev", p->orig_iface, NULL};
-        (void)run_ip_cmd(d5);
-    }
-    p->routing_configured = 0;
-    LOG_INF("split tunnel routes cleaned up");
-}
-
-/* ================================================================
- *  Kill switch setup / cleanup
- * ================================================================ */
-
-static void cleanup_killswitch(platform_ctx_t *p);
-
-static int
-setup_killswitch(platform_ctx_t *p)
-{
-    if (!p->killswitch_enabled || p->killswitch_active) return 0;
-
-    snprintf(p->ks_comment, sizeof(p->ks_comment), "mqvpn-ks:%d", (int)getpid());
-    int is_v6 = (p->server_addr.ss_family == AF_INET6);
-
-    /* IPv4 rules (always needed) */
-    const char *at[] = {"iptables", "-I", "OUTPUT", "-o", p->tun.name,
-                        "-j", "ACCEPT", "-m", "comment", "--comment", p->ks_comment, NULL};
-    const char *al[] = {"iptables", "-I", "OUTPUT", "-o", "lo",
-                        "-j", "ACCEPT", "-m", "comment", "--comment", p->ks_comment, NULL};
-    const char *da[] = {"iptables", "-A", "OUTPUT", "-j", "DROP",
-                        "-m", "comment", "--comment", p->ks_comment, NULL};
-
-    if (run_iptables_cmd(at) < 0 || run_iptables_cmd(al) < 0 || run_iptables_cmd(da) < 0) {
-        LOG_WRN("failed to set up iptables kill switch rules");
-        p->killswitch_active = 1;
-        cleanup_killswitch(p);
-        return -1;
-    }
-    p->killswitch_active = 1;
-
-    int pfx = mqvpn_sa_host_prefix(&p->server_addr);
-    char sc[INET6_ADDRSTRLEN + 5];
-    snprintf(sc, sizeof(sc), "%s/%d", p->server_ip_str, pfx);
-    char ps[8];
-    snprintf(ps, sizeof(ps), "%d", p->server_port);
-
-    if (!is_v6) {
-        const char *as[] = {"iptables", "-I", "OUTPUT", "-p", "udp",
-                            "-d", sc, "--dport", ps, "-o", p->orig_iface,
-                            "-j", "ACCEPT", "-m", "comment", "--comment", p->ks_comment, NULL};
-        if (run_iptables_cmd(as) < 0) { cleanup_killswitch(p); return -1; }
-    } else {
-        const char *v6s[] = {"ip6tables", "-I", "OUTPUT", "-p", "udp",
-                             "-d", sc, "--dport", ps, "-o", p->orig_iface,
-                             "-j", "ACCEPT", "-m", "comment", "--comment", p->ks_comment, NULL};
-        const char *v6l[] = {"ip6tables", "-I", "OUTPUT", "-o", "lo",
-                             "-j", "ACCEPT", "-m", "comment", "--comment", p->ks_comment, NULL};
-        const char *v6d[] = {"ip6tables", "-A", "OUTPUT", "-j", "DROP",
-                             "-m", "comment", "--comment", p->ks_comment, NULL};
-        if (run_iptables_cmd(v6s) < 0 || run_iptables_cmd(v6l) < 0 ||
-            run_iptables_cmd(v6d) < 0) {
-            cleanup_killswitch(p);
-            return -1;
-        }
-    }
-
-    /* IPv6 data plane TUN rule */
-    if (p->has_v6) {
-        const char *v6t[] = {"ip6tables", "-I", "OUTPUT", "-o", p->tun.name,
-                             "-j", "ACCEPT", "-m", "comment", "--comment", p->ks_comment, NULL};
-        (void)run_iptables_cmd(v6t);
-        if (!is_v6) {
-            const char *v6l[] = {"ip6tables", "-I", "OUTPUT", "-o", "lo",
-                                 "-j", "ACCEPT", "-m", "comment", "--comment", p->ks_comment, NULL};
-            const char *v6d[] = {"ip6tables", "-A", "OUTPUT", "-j", "DROP",
-                                 "-m", "comment", "--comment", p->ks_comment, NULL};
-            (void)run_iptables_cmd(v6l);
-            (void)run_iptables_cmd(v6d);
-        }
-    }
-    LOG_INF("kill switch enabled (comment=%s)", p->ks_comment);
-    return 0;
-}
-
-static void
-cleanup_killswitch(platform_ctx_t *p)
-{
-    if (!p->killswitch_active) return;
-
-    int is_v6 = (p->server_addr.ss_family == AF_INET6);
-    int pfx = mqvpn_sa_host_prefix(&p->server_addr);
-    char sc[INET6_ADDRSTRLEN + 5];
-    snprintf(sc, sizeof(sc), "%s/%d", p->server_ip_str, pfx);
-    char ps[8];
-    snprintf(ps, sizeof(ps), "%d", p->server_port);
-
-    /* Remove IPv4 rules */
-    const char *dt[] = {"iptables", "-D", "OUTPUT", "-o", p->tun.name,
-                        "-j", "ACCEPT", "-m", "comment", "--comment", p->ks_comment, NULL};
-    while (run_iptables_cmd(dt) == 0) {}
-    const char *dl[] = {"iptables", "-D", "OUTPUT", "-o", "lo",
-                        "-j", "ACCEPT", "-m", "comment", "--comment", p->ks_comment, NULL};
-    while (run_iptables_cmd(dl) == 0) {}
-    const char *dd[] = {"iptables", "-D", "OUTPUT", "-j", "DROP",
-                        "-m", "comment", "--comment", p->ks_comment, NULL};
-    while (run_iptables_cmd(dd) == 0) {}
-
-    if (!is_v6) {
-        const char *ds[] = {"iptables", "-D", "OUTPUT", "-p", "udp",
-                            "-d", sc, "--dport", ps, "-o", p->orig_iface,
-                            "-j", "ACCEPT", "-m", "comment", "--comment", p->ks_comment, NULL};
-        while (run_iptables_cmd(ds) == 0) {}
-    } else {
-        const char *v6s[] = {"ip6tables", "-D", "OUTPUT", "-p", "udp",
-                             "-d", sc, "--dport", ps, "-o", p->orig_iface,
-                             "-j", "ACCEPT", "-m", "comment", "--comment", p->ks_comment, NULL};
-        while (run_iptables_cmd(v6s) == 0) {}
-        const char *v6l[] = {"ip6tables", "-D", "OUTPUT", "-o", "lo",
-                             "-j", "ACCEPT", "-m", "comment", "--comment", p->ks_comment, NULL};
-        while (run_iptables_cmd(v6l) == 0) {}
-        const char *v6d[] = {"ip6tables", "-D", "OUTPUT", "-j", "DROP",
-                             "-m", "comment", "--comment", p->ks_comment, NULL};
-        while (run_iptables_cmd(v6d) == 0) {}
-    }
-
-    /* IPv6 data plane TUN rule */
-    const char *v6dt[] = {"ip6tables", "-D", "OUTPUT", "-o", p->tun.name,
-                          "-j", "ACCEPT", "-m", "comment", "--comment", p->ks_comment, NULL};
-    while (run_iptables_cmd(v6dt) == 0) {}
-
-    if (!is_v6) {
-        const char *v6l2[] = {"ip6tables", "-D", "OUTPUT", "-o", "lo",
-                              "-j", "ACCEPT", "-m", "comment", "--comment", p->ks_comment, NULL};
-        while (run_iptables_cmd(v6l2) == 0) {}
-        const char *v6d2[] = {"ip6tables", "-D", "OUTPUT", "-j", "DROP",
-                              "-m", "comment", "--comment", p->ks_comment, NULL};
-        while (run_iptables_cmd(v6d2) == 0) {}
-    }
-    p->killswitch_active = 0;
-    LOG_INF("kill switch rules removed");
-}
 
 /* ================================================================
  *  libmqvpn callbacks
